Refactor: extract the business impact analysis
